ELMIS-593: fix the adjustment summary report

+++ modules/report/src/main/java/org/openlmis/report/model/report/AdjustmentSummaryReport.java
@@ -26,6 +26,6 @@ public class AdjustmentSummaryReport implements ReportData {
     private String facilityName;
     private String supplyingFacility;
     private String adjustmentType;
-    private Double adjustment;
+    private int adjustment;
 
 }
